# app/Http/Controllers/MeasureUnitController.php
<?php

namespace App\Http\Controllers;

use App\DTOs\MeasureUnitDto;
use App\Http\Requests\MeasureUnitRequest;
use App\Services\MeasureUnitService;
use Exception;

use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;
use Illuminate\Support\Facades\Log;

use App\Models\MeasureUnit;

class MeasureUnitController extends Controller
{
    private $measureUnitService;

    public function __construct(MeasureUnitService $measureUnitService) {
        $this->measureUnitService = $measureUnitService;
    }

    /**
     * @OA\Post(
     *     path="/measure-units",
     *     summary="Create a measure unit",
     *     tags={"Measure Units"},
     *     security={{"bearerAuth":{}}},
     *     @OA\RequestBody(
     *          required=true,
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitRequest")
     *     ),
     *     @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitResponse")
     *     ),
     *     @OA\Response(response=400, description="Invalid request")
     * )
     */
    public function createUnit(MeasureUnitRequest $request) {
        $measureUnit = MeasureUnitDto::fromRequest($request);
        $measureUnit = $this->measureUnitService->createMeasureUnit($measureUnit);
        return response()->json($measureUnit, 200);
    }

    /**
     * @OA\Get(
     *     path="/measure-units",
     *     summary="Get list of measure units",
     *     tags={"Measure Units"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Response(
     *          response=200,
     *          description="Success",
     *          @OA\JsonContent(
     *              type="array",
     *              @OA\Items(ref="#/components/schemas/MeasureUnitResponse")
     *          )
     *     )
     * )
     */
    public function getAllMeasureUnits() {
        $categories = MeasureUnit::get();
        return response()->json($categories, 200);
    }

    /**
     * @OA\Get(
     *     path="/measure-units/{id}",
     *     summary="Get measure unit by id",
     *     tags={"Measure Units"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="Id of the measure unit",
     *         @OA\Schema(type="integer"),
     *         required=true,
     *         example=1
     *     ),
     *     @OA\Response(
     *          response=200,
     *          description="Success",
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitResponse")
     *     ),
     *     @OA\Response(response=400, description="Invalid measure unit id")
     * )
     */
    public function getMeasureUnit($id) {
        $measureUnit = MeasureUnit::find('id', $id);
        return response()->json($measureUnit, 200);
    }

    /**
     * @OA\Put(
     *     path="/measure-units/{id}",
     *     summary="Update a measure unit",
     *     tags={"Measure Units"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="Id of the measure unit",
     *         @OA\Schema(type="integer"),
     *         required=true,
     *         example=1
     *     ),
     *     @OA\RequestBody(
     *          required=true,
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitRequest")
     *     ),
     *     @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitResponse")
     *     ),
     *     @OA\Response(response=400, description="Invalid request"),
     *     @OA\Response(response=500, description="Server error")
     * )
     */
    public function updateMeasureUnit($id, Request $request) {
        try {
            $measureUnit = MeasureUnit::find($id);

            if ($measureUnit == null) {
            return response()->json(["code"=>"Invalid Input", "message"=>"Measure unit not found"], 400);
            }

            $measureUnit->unit_name = $request->unit_name;
            $measureUnit->description = $request->description;
            $measureUnit->active = $request->active;
            $measureUnit->update();

            return response()->json($measureUnit, 200);
        } catch(Exception $ex) {
            Log::channel('debug')->error($ex->getMessage());
            return response()->json(["message"=>$ex->getMessage()], 500);
        }
    }

    /**
     * @OA\Delete(
     *     path="/measure-units/{id}",
     *     summary="Delete measure unit",
     *     tags={"Measure Units"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="Id of the measure unit",
     *         @OA\Schema(type="integer"),
     *         required=true,
     *         example=1
     *     ),
     *     @OA\Response(
     *          response=200,
     *          description="Success",
     *          @OA\JsonContent(ref="#/components/schemas/MeasureUnitResponse")
     *     ),
     *     @OA\Response(response=400, description="Invalid measure unit id")
     * )
     */
    public function deleteMeasureUnit($id) {
        $measureUnit = MeasureUnit::find($id);

        if ($measureUnit == null) {
            return response()->json(["code"=>"Invalid Input", "message"=>"Measure unit not found"], 400);
        }

        $measureUnit->active = 0;
        $measureUnit->update();

        return response()->json($measureUnit, 200);
    }
}