# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dreamers Solutions ERP APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/measure-units": {
            "get": {
                "tags": [
                    "Measure Units"
                ],
                "summary": "Get list of measure units",
                "operationId": "6931d7cff070feab46920b5f44014a2a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MeasureUnitResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Measure Units"
                ],
                "summary": "Create a measure unit",
                "operationId": "ef3be914ad164b98cb8c1f982bb60966",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasureUnitRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasureUnitResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/measure-units/{id}": {
            "get": {
                "tags": [
                    "Measure Units"
                ],
                "summary": "Get measure unit by id",
                "operationId": "2b7529e930b3361af196f06fb3b6a8c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the measure unit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasureUnitResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid measure unit id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Measure Units"
                ],
                "summary": "Update a measure unit",
                "operationId": "c1bcaa798161ec9e0e5fe92d2eae65f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the measure unit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasureUnitRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasureUnitResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Measure Units"
                ],
                "summary": "Delete measure unit",
                "operationId": "fae42fab6c28f36e44f4ac4b62b6f36c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the measure unit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasureUnitResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid measure unit id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/permissions/{locationId}/{moduleId}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions for menu by location and module",
                "operationId": "a8fbd1036cc62a063ec65ab800ac5737",
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "Location id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "moduleId",
                        "in": "path",
                        "description": "Module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PermissionResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MeasureUnitRequest": {
                "title": "MeasureUnitRequest",
                "description": "Measure unit request body",
                "required": [
                    "unitName"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the measure unit",
                        "type": "integer"
                    },
                    "unitName": {
                        "description": "Name of the measure unit",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the measure unit",
                        "type": "string"
                    },
                    "active": {
                        "description": "Active status of the measure unit",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "MeasureUnitResponse": {
                "title": "MeasureUnitResponse",
                "description": "Measure unit response model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the measure unit",
                        "type": "integer"
                    },
                    "unit_name": {
                        "description": "Name of the measure unit",
                        "type": "string"
                    },
                    "unit_code": {
                        "description": "Code of the measure unit",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the measure unit",
                        "type": "string"
                    },
                    "active": {
                        "description": "Active status of the measure unit",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "PermissionResponse": {
                "title": "PermissionResponse",
                "description": "Permission record",
                "properties": {
                    "id": {
                        "description": "Permission id",
                        "type": "integer"
                    },
                    "permission_code": {
                        "description": "Permission code",
                        "type": "string"
                    },
                    "fxml_path": {
                        "description": "FXML path for menu permission",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "Display name",
                        "type": "string"
                    },
                    "parent_id": {
                        "description": "Parent permission id",
                        "type": "integer"
                    },
                    "depth": {
                        "description": "Permission depth in hierarchy",
                        "type": "integer"
                    },
                    "permission_type": {
                        "description": "Permission type",
                        "type": "string"
                    },
                    "active": {
                        "description": "Active status",
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from the login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}